relaese : 서버 파일 완료

+++ src/main/java/com/project/savingbee/productCompare/util/CalcEngine.java
@@ -7,11 +7,13 @@ import lombok.Getter;
 
 // 세후 이자, 실수령액 계산
 public final class CalcEngine {
+
   /**
    * 예금 만기시 실수령액 계산
-   * @param amount 예치금액
-   * @param rate 세후 이자율
-   * @param termMonth 예치기간
+   *
+   * @param amount       예치금액
+   * @param rate         세후 이자율
+   * @param termMonth    예치기간
    * @param intrRateType 이자계산방식(단리/복리)
    */
   public static CalcResult deposit(
@@ -23,20 +25,21 @@ public final class CalcEngine {
 
     BigDecimal interest = intrRateType.equals("S")
         ? A.multiply(R).multiply(BigDecimal.valueOf(termMonth)
-            .divide(BigDecimal.valueOf(12), 10, RoundingMode.HALF_UP))
+        .divide(BigDecimal.valueOf(12), 10, RoundingMode.HALF_UP))
         : A.multiply(BigDecimal.ONE.add(I).pow(termMonth).subtract(BigDecimal.ONE));
 
-    BigDecimal afterTaxInterest = interest.setScale(0, RoundingMode.DOWN);
-    long amountReceived = A.longValue() + afterTaxInterest.longValue();
+    long afterTaxInterest = interest.setScale(0, RoundingMode.DOWN).longValue();
+    long amountReceived = A.longValue() + afterTaxInterest;
 
     return new CalcResult(afterTaxInterest, amountReceived);
   }
 
   /**
    * 적금 만기시 실수령액 계산
-   * @param amount 월 납입금액
-   * @param rate 세후 이자율
-   * @param termMonth 예치기간
+   *
+   * @param amount       월 납입금액
+   * @param rate         세후 이자율
+   * @param termMonth    예치기간
    * @param intrRateType 이자계산방식(단리/복리)
    */
   public static CalcResult savings(
@@ -65,9 +68,9 @@ public final class CalcEngine {
       }
     }
 
-    BigDecimal afterTaxInterest = interest.setScale(0, RoundingMode.DOWN);
+    long afterTaxInterest = interest.setScale(0, RoundingMode.DOWN).longValue();
     long amountReceived =
-        totalAmount.setScale(0, RoundingMode.DOWN).longValue() + afterTaxInterest.longValue();
+        totalAmount.setScale(0, RoundingMode.DOWN).longValue() + afterTaxInterest;
 
     return new CalcResult(afterTaxInterest, amountReceived);
   }
@@ -75,7 +78,8 @@ public final class CalcEngine {
   @Getter
   @AllArgsConstructor
   public static class CalcResult {
-    private final BigDecimal afterTaxInterest;  // 세후 이자
+
+    private final long afterTaxInterest;  // 세후 이자
     private final long amountReceived;  // 실수령액
   }
